day8. chap6 과제

+++ src/day8/chap6_OOP1/EmployeeProj/Company.java
@@ -40,8 +40,8 @@ public class Company {
         }
     }
     // 전체 급여 조회
-    int getTotalPay(){
-        int total = 0;
+    long getTotalPay(){
+        long total = 0;
         for(int i = 0; i<empCnt; i++){
             total += emps[i].pay;
         }
